requirements.txt
alembic
python-dotenv
pydantic==2.12.5
pydantic-settings
pytest
httpx
pytest-asyncio
fastapi==0.102.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
fastapi-cache2==0.2.2
slowapi==0.1.5